add types of metadata and some unit tests
