+++ packages/metadata/README.md
@@ -12,6 +12,14 @@ npm install @bosonprotocol/metadata
 yarn add @bosonprotocol/metadata
 ```
 
-## Usage
+## Details
 
-> TODO
+Several Metadata Schema are defined, for different purpose:
+- SELLER: Metadata Schema for a Seller
+- COLLECTION: Metadata Schema for a Collection of Offers
+- rNFT: Metadata Schema for a Redeemable NFT (similar to BASE)
+- BASE: Base Schema Metadata for an Offer
+- PRODUCT_V1: Metadata Schema for a Physical Product Offer (version 1)
+- BUNDLE: Metadata Schema for a Bundle Offer
+- ITEM_PRODUCT_V1: Metadata Schema for an Physical Product Item (referenced in a Bundle Offer)
+- ITEM_NFT:  Metadata Schema for an NFT Item (referenced in a Bundle Offer)
