+++ packages/metadata/src/base/index.ts
@@ -1,16 +1,10 @@
 import { buildYup } from "schema-to-yup";
 import { SchemaOf } from "yup";
 import schema from "./schema.json";
+import { RNftMetadata } from "../rNFT";
 
 export const baseMetadataSchema: SchemaOf<BaseMetadata> = buildYup(schema, {});
 
-export type BaseMetadata = {
-  schemaUrl: string;
+export type BaseMetadata = Omit<RNftMetadata, "type"> & {
   type: "BASE";
-  name: string;
-  description: string;
-  externalUrl: string;
-  licenseUrl: string;
-  condition?: string;
-  animationUrl?: string;
 };
