+++ packages/metadata/src/bundle/schema.json
@@ -2,7 +2,7 @@
   "type": "object",
   "properties": {
     "schemaUrl": { "type": "string" },
-    "type": { "type": "string", "enum": ["BASE"] },
+    "type": { "type": "string", "enum": ["BUNDLE"] },
     "name": { "type": "string" },
     "description": { "type": "string" },
     "image": { "type": "string" },
@@ -21,7 +21,18 @@
         },
         "required": ["trait_type", "value"]
       }
+    },
+    "items": {
+      "type": "array",
+      "items": {
+        "type": "object",
+        "properties": {
+          "url": { "type": "string" }
+        },
+        "required": ["url"]
+      },
+      "minItems": 1
     }
   },
-  "required": ["schemaUrl", "type", "name", "description", "externalUrl", "licenseUrl"]
+  "required": ["schemaUrl", "type", "name", "description", "externalUrl", "licenseUrl", "items"]
 }
