+++ packages/metadata/tests/index.test.ts
@@ -1,242 +1,21 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { validateMetadata, AnyMetadata, productV1 } from "../src/index";
+import {
+  validateMetadata,
+  AnyMetadata,
+  productV1,
+  productV1Item,
+  nftItem
+} from "../src/index";
 import productV1ValidFullOffer from "./product-v1/valid/fullOffer.json";
 import productV1ValidMinimalOffer from "./product-v1/valid/minimalOffer.json";
+import { productMissingArguments as productV1_productMissingArguments } from "./product-v1/missingArguments";
+import productV1ItemValidMinimal from "./product-v1-item/valid/minimal.json";
+import productV1ItemValidFull from "./product-v1-item/valid/full.json";
+import { productMissingArguments as productV1Item_productMissingArguments } from "./product-v1-item/missingArguments";
+import nftItemValidMinimal from "./nft-item/valid/minimal.json";
+import nftItemValidFull from "./nft-item/valid/full.json";
+import { missingArguments as nftItemMissingArguments } from "./nft-item/missingArguments";
 import cloneDeep from "clone-deep";
-
-const productMissingArguments = [
-  {
-    arg: "schemaUrl",
-    data: { schemaUrl: undefined },
-    error: /schemaUrl is a required field/
-  },
-  {
-    arg: "type",
-    data: { type: undefined },
-    error: /Metadata validation failed for unknown type/
-  },
-  {
-    arg: "uuid",
-    data: { uuid: undefined },
-    error: /uuid is a required field/
-  },
-  {
-    arg: "name",
-    data: { name: undefined },
-    error: /name is a required field/
-  },
-  {
-    arg: "description",
-    data: { description: undefined },
-    error: /description is a required field/
-  },
-  {
-    arg: "externalUrl",
-    data: { externalUrl: undefined },
-    error: /externalUrl is a required field/
-  },
-  {
-    arg: "animationUrl",
-    data: { animationUrl: undefined },
-    error: undefined // don't check the error message
-  },
-  {
-    arg: "licenseUrl",
-    data: { licenseUrl: undefined },
-    error: /licenseUrl is a required field/
-  },
-  {
-    arg: "image",
-    data: { image: undefined },
-    error: /image is a required field/
-  },
-  {
-    arg: "attributes",
-    data: { attributes: undefined },
-    error: /attributes is a required field/
-  },
-  {
-    arg: "product",
-    data: { product: undefined },
-    error: undefined // don't check the error message
-  },
-  {
-    arg: "shipping",
-    data: { shipping: undefined },
-    error: undefined // don't check the error message
-  },
-  {
-    arg: "shipping.returnPeriod",
-    data: { shipping: { returnPeriod: undefined } },
-    error: /shipping.returnPeriod is a required field/
-  },
-  {
-    arg: "exchangePolicy",
-    data: { exchangePolicy: undefined },
-    error: undefined // don't check the error message
-  },
-  {
-    arg: "exchangePolicy.uuid",
-    data: { exchangePolicy: { uuid: undefined } },
-    error: /exchangePolicy.uuid is a required field/
-  },
-  {
-    arg: "exchangePolicy.version",
-    data: { exchangePolicy: { version: undefined } },
-    error: /exchangePolicy.version is a required field/
-  },
-  {
-    arg: "exchangePolicy.version not a number",
-    data: { exchangePolicy: { version: "not_a_number" } },
-    // eslint-disable-next-line no-useless-escape
-    error: /exchangePolicy.version must be a \`number\` type/
-  },
-  {
-    arg: "exchangePolicy.template",
-    data: { exchangePolicy: { template: undefined } },
-    error: /exchangePolicy.template is a required field/
-  },
-  {
-    arg: "exchangePolicy.sellerContactMethod",
-    data: { exchangePolicy: { sellerContactMethod: undefined } },
-    error: /exchangePolicy.sellerContactMethod is a required field/
-  },
-  {
-    arg: "exchangePolicy.disputeResolverContactMethod",
-    data: { exchangePolicy: { disputeResolverContactMethod: undefined } },
-    error: /exchangePolicy.disputeResolverContactMethod is a required field/
-  },
-  {
-    arg: "product.uuid",
-    data: { product: { uuid: undefined } },
-    error: /product.uuid is a required field/
-  },
-  {
-    arg: "product.version",
-    data: { product: { version: undefined } },
-    error: /product.version is a required field/
-  },
-  {
-    arg: "product.version not a number",
-    data: { product: { version: "not_a_number" } },
-    // eslint-disable-next-line no-useless-escape
-    error: /product.version must be a \`number\` type/
-  },
-  {
-    arg: "product.title",
-    data: { product: { title: undefined } },
-    error: /product.title is a required field/
-  },
-  {
-    arg: "product.description",
-    data: { product: { description: undefined } },
-    error: /product.description is a required field/
-  },
-  {
-    arg: "product.productionInformation_brandName",
-    data: { product: { productionInformation_brandName: undefined } },
-    error: /product.productionInformation_brandName is a required field/
-  },
-  {
-    arg: "product.details_offerCategory",
-    data: { product: { details_offerCategory: undefined } },
-    error: /product.details_offerCategory is a required field/
-  },
-  {
-    arg: "product.visuals_images",
-    data: { product: { visuals_images: undefined } },
-    error: /product.visuals_images is a required field/
-  },
-  {
-    arg: "variations[x].type",
-    data: { variations: [{ option: "anOption" }] },
-    error: /variations\[0\].type is a required field/
-  },
-  {
-    arg: "variations[x].option",
-    data: { variations: [{ type: "aType" }] },
-    error: /variations\[0\].option is a required field/
-  },
-  {
-    arg: "seller.images[x].url",
-    data: { seller: { images: [{ tag: "aTag" }] } },
-    error: /seller.images\[0\].url is a required field/
-  },
-  {
-    arg: "seller",
-    data: { seller: undefined },
-    error: undefined // don't check the error message
-  },
-  {
-    arg: "seller.defaultVersion",
-    data: { seller: { defaultVersion: undefined } },
-    error: /seller.defaultVersion is a required field/
-  },
-  {
-    arg: "seller.contactLinks",
-    data: { seller: { contactLinks: undefined } },
-    error: /seller.contactLinks is a required field/
-  },
-  {
-    arg: "seller.name",
-    data: { seller: { name: undefined } },
-    error: /seller.name is a required field/
-  },
-  // {
-  //   arg: "seller.contactLinks empty array",
-  //   data: { seller: { defaultVersion: 1, contactLinks: [] } },
-  //   error: undefined // don't check the error message
-  // }, // it should fail because minItems is set to 1 (does not work as expected...)
-  {
-    arg: "seller.contactLinks[x].url",
-    data: { seller: { contactLinks: [{ tag: "aTag" }] } },
-    error: /seller.contactLinks\[0\].url is a required field/
-  },
-  {
-    arg: "seller.contactLinks[x].tag",
-    data: { seller: { contactLinks: [{ url: "example.com" }] } },
-    error: /seller.contactLinks\[0\].tag is a required field/
-  },
-  {
-    arg: "shipping.supportedJurisdictions[x].label",
-    data: {
-      shipping: { supportedJurisdictions: [{ deliveryTime: "very soon" }] }
-    },
-    error: /shipping.supportedJurisdictions\[0\].label is a required field/
-  },
-  {
-    arg: "shipping.supportedJurisdictions[x].deliveryTime",
-    data: {
-      shipping: { supportedJurisdictions: [{ label: "aLabel" }] }
-    },
-    error:
-      /shipping.supportedJurisdictions\[0\].deliveryTime is a required field/
-  }
-  // {
-  //   arg: "attributes",
-  //   data: {
-  //     attributes: [
-  //       {
-  //         trait_type: "Offer Category",
-  //         value: "PHYSICAL"
-  //       }
-  //     ]
-  //   },
-  //   error: undefined // don't check the error message
-  // } // it should fail because minItems is set to 2 (does not work as expected...)
-
-  // {
-  //   arg: "product.visuals.images",
-  //   data: { product: { visuals: { images: [] } } },
-  //   error: undefined // don't check the error message
-  // } // it should fail because minItems is set to 1 (does not work as expected...)
-
-  // {
-  //   arg: "seller",
-  //   data: { seller: undefined },
-  //   error: /seller is a required field/
-  // } // it should fail because seller is a required field (not as expected)
-];
 
 describe("#validateMetadata()", () => {
   test("throw for invalid type", () => {
@@ -320,7 +99,7 @@ describe("#validateMetadata()", () => {
       );
     });
 
-    test.each(productMissingArguments)(
+    test.each(productV1_productMissingArguments)(
       "throw for missing argument '$arg'",
       ({ data, error }) => {
         const product = (data as any).product
@@ -389,5 +168,116 @@ describe("#validateMetadata()", () => {
         } as any as AnyMetadata)
       ).toThrow();
     });
+  });
+
+  describe("PRODUCT_V1_ITEM", () => {
+    test("throw for invalid object", () => {
+      expect(() =>
+        validateMetadata({
+          type: "PRODUCT_V1_ITEM"
+        } as unknown as AnyMetadata)
+      ).toThrow();
+    });
+    test("not throw for full metadata schema", () => {
+      expect(
+        validateMetadata(productV1ItemValidFull as unknown as AnyMetadata)
+      ).toBeTruthy();
+    });
+    test("not throw for minimal metadata schema", () => {
+      expect(
+        validateMetadata(productV1ItemValidMinimal as unknown as AnyMetadata)
+      ).toBeTruthy();
+    });
+    test("throw for too long value", () => {
+      const metadata = cloneDeep(
+        productV1ItemValidFull as unknown as productV1Item.ProductV1Item
+      );
+
+      const imageIndex = metadata.product.visuals_images.length;
+      metadata.product.visuals_images.push({
+        url: new Array(10000).join(",")
+      });
+      expect(() => validateMetadata(metadata)).toThrow(
+        `Key product.visuals_images.${imageIndex}.url of metadata exceeds 2048 characters`
+      );
+    });
+    test.each(productV1Item_productMissingArguments)(
+      "throw for missing argument '$arg'",
+      ({ data, error }) => {
+        const product = (data as any).product
+          ? {
+              ...productV1ValidFullOffer,
+              ...data,
+              product: {
+                ...productV1ValidFullOffer.product,
+                ...(data as any).product
+              }
+            }
+          : (data as any).exchangePolicy
+          ? {
+              ...productV1ValidFullOffer,
+              ...data,
+              exchangePolicy: {
+                ...productV1ValidFullOffer.exchangePolicy,
+                ...(data as any).exchangePolicy
+              }
+            }
+          : {
+              ...productV1ValidFullOffer,
+              ...data
+            };
+        const result = expect(() =>
+          validateMetadata(product as any as AnyMetadata)
+        );
+        if (error) {
+          result.toThrow(error);
+        }
+      }
+    );
+  });
+
+  describe("NFT_ITEM", () => {
+    test("throw for invalid object", () => {
+      expect(() =>
+        validateMetadata({
+          type: "ITEM_NFT"
+        } as unknown as AnyMetadata)
+      ).toThrow();
+    });
+    test("not throw for full metadata schema", () => {
+      expect(
+        validateMetadata(nftItemValidFull as unknown as AnyMetadata)
+      ).toBeTruthy();
+    });
+    test("not throw for minimal metadata schema", () => {
+      expect(
+        validateMetadata(nftItemValidMinimal as unknown as AnyMetadata)
+      ).toBeTruthy();
+    });
+    test("throw for too long value", () => {
+      const metadata = cloneDeep(
+        nftItemValidFull as unknown as nftItem.NftItem
+      ) as nftItem.NftItem;
+
+      metadata.animationUrl = new Array(10000).join(",");
+      expect(() => validateMetadata(metadata)).toThrow(
+        `Key animationUrl of metadata exceeds 2048 characters`
+      );
+    });
+    test.each(nftItemMissingArguments)(
+      "throw for missing argument '$arg'",
+      ({ data, error }) => {
+        const metadata = {
+          ...nftItemValidFull,
+          ...data
+        };
+        const result = expect(() =>
+          validateMetadata(metadata as any as AnyMetadata)
+        );
+        if (error) {
+          result.toThrow(error);
+        }
+      }
+    );
   });
 });
