+++ packages/metadata/tests/nft-item/valid/full.json
@@ -1,0 +1,24 @@
+{
+  "schemaUrl": "https://json-schema.org",
+  "type": "ITEM_NFT",
+  "name": "Boson X MetaFactory Hoodie",
+  "description": "The future is here and it starts with a limited edition MetaFactory X Boson Protocol Hoodie... This unisex hoodie, in a metablack coloureway with reflective stripes, is made from a heavyweight French terry cotton. It has an adjustable hood and full colour interior, elasticated trims, reflective lining, and reflective back panels... Nothing says open metaverse more than this collab between MetaFactory and Boson Protocol. Limited Edition: Only 75 ever made.",
+  "image": "https://bsn-portal-production-image-upload-storage.s3.amazonaws.com/fc11acc4-e27b-4ede-b7d3-f16735fab406",
+  "externalUrl": "https://app.bosonportal.io/#/offer-uuid/ecf2a6dc-555b-41b5-aca8-b7e29eebbb30",
+  "animationUrl": "https://app.bosonportal.io/animationUrl",
+  "attributes": [
+    {
+      "trait_type": "Creator",
+      "value": "MetaFactory"
+    },
+    {
+      "trait_type": "Redeemable Until",
+      "display_type": "date",
+      "value": "1658492054"
+    },
+    {
+      "trait_type": "Offer Category",
+      "value": "PHYGITAL"
+    }
+  ]
+}
