+++ packages/metadata/tests/nft-item/valid/minimal.json
@@ -1,0 +1,5 @@
+{
+  "schemaUrl": "https://json-schema.org",
+  "type": "ITEM_NFT",
+  "name": "Boson X MetaFactory Hoodie"
+}
