+++ packages/metadata/tests/product-v1-item/valid/minimal.json
@@ -1,0 +1,31 @@
+{
+  "schemaUrl": "https://json-schema.org",
+  "type": "ITEM_PRODUCT_V1",
+  "product": {
+    "uuid": "77593bb2-f797-11ec-b939-0242ac120002",
+    "version": 1,
+    "title": "Boson X MetaFactory Hoodie",
+    "description": "The future is here and it starts with a limited edition MetaFactory X Boson Protocol Hoodie... This unisex hoodie, in a metablack coloureway with reflective stripes, is made from a heavyweight French terry cotton. It has an adjustable hood and full colour interior, elasticated trims, reflective lining, and reflective back panels... Nothing says open metaverse more than this collab between MetaFactory and Boson Protocol. Limited Edition: Only 75 ever made.",
+    "productionInformation_brandName": "Boson X MetaFactory",
+    "details_offerCategory": "PHYSICAL",
+    "visuals_images": [
+      {
+        "url": "https://bsn-portal-production-image-upload-storage.s3.amazonaws.com/5ba6461b-e4b4-444d-90bf-0c9de1835c35",
+        "tag": "product_image",
+        "height": 2736,
+        "width": 2310,
+        "type": "image/jpeg"
+      }
+    ]
+  },
+  "shipping": {
+    "returnPeriod": "14"
+  },
+  "exchangePolicy": {
+    "uuid": "1e82f742-5639-427f-87b2-b91a7e51e40a",
+    "version": 1,
+    "template": "https://github.com/dr-harper/example-rmd-templates/blob/master/inst/rmarkdown/templates/01-minimal/template.yaml",
+    "sellerContactMethod": "DAPP_CHAT",
+    "disputeResolverContactMethod": "EMAIL"
+  }
+}
